requirements.txt
aiohttp==3.8.3
pandas==1.1.5
PyJWT==2.4.0
requests==2.18.4
SQLAlchemy==1.4.44
psycopg2-binary==2.9.5
psycopg2-pool==1.1